requirements.txt
requests
pandas
numpy
ta
python-telegram-bot